add connection guide (#22)
* update bucket name

* build .zip file for ci/cd

* remove is_nullable

* update submit_job

* update external table

* update external table bugs

* update external table bugs

* fix external tables bugs

* fix external tables bugs

* fix external tables bugs

* update external schema

* update bigquery

* add Partition Columns

* update bigquery

* update uri

* update partitioned table

* update landing to bronze

* done for elt source to silver staging

* black format code

* lint format code

* update pyspark types

* add libs

* fix data types

* add connection guide

---------

+++ pyspark/elt/libs/aw/schema.py
@@ -28,7 +28,7 @@ MYSQL_TO_SPARK_DATA_TYPE = {
     "int": "IntegerType",         
     "integer": "IntegerType",     
     "bigint": "LongType",         
-    "float": "FloatType",         
+    "float": "DoubleType",         
     "double": "DoubleType",       
     "decimal": "DecimalType",     
     "numeric": "DecimalType",
